Outsourced disk based methods
