+++ src/pymor/discretizers/elliptic.py
@@ -3,13 +3,9 @@
 # License: BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
 #
 # Contributors: lucas-ca <lucascamp@web.de>
-#		Falk Meyer <falk.meyer@wwu.de>
 
 from __future__ import absolute_import, division, print_function
-from scipy import io
 
-import numpy as np
-import ConfigParser
 
 from pymor.analyticalproblems.elliptic import EllipticProblem
 from pymor.discretizations.basic import StationaryDiscretization
@@ -21,9 +17,7 @@ from pymor.grids.tria import TriaGrid
 from pymor.gui.qt import PatchVisualizer, Matplotlib1DVisualizer
 from pymor.operators import cg, fv
 from pymor.operators.constructions import LincombOperator
-from pymor.operators.numpy import NumpyMatrixOperator
-from pymor.parameters.functionals import ExpressionParameterFunctional
-from pymor.parameters.spaces import CubicParameterSpace
+
 
 def discretize_elliptic_cg(analytical_problem, diameter=None, domain_discretizer=None,
                            grid=None, boundary_info=None):
@@ -204,127 +198,3 @@ def discretize_elliptic_fv(analytical_problem, diameter=None, domain_discretizer
                                               parameter_space=parameter_space, name='{}_FV'.format(p.name))
 
     return discretization, {'grid': grid, 'boundary_info': boundary_info}
-
-def discretize_elliptic_disk(parameter_file):
-    """Generates stationary discretization only based on system relevant data given as .mat files on the hard disc. The path and further
-     specifications to these objects are given in an '.ini' parameter file (see example below). Suitable for discrete problems given by::
-
-	 L(u(w), w) = F(w)
-
-    with an operator L and a linear functional F with a parameter w  given as stiffness matrices and rhs vectors in an affine decomposition 
-    on the hard disc.
-
-    Parameters
-    ----------
-    parameterFile
-	    String containing the path to the .ini parameterfile.
-
-    Returns
-    -------
-    discretization
-        The |Discretization| that has been generated.
-
-
-    Example parameter_file
-    -------
-    Following parameter file is suitable for a discrete elliptic problem with
-
-    L(u(w), w) = (f_1(w)*K1 + f_2(w)*K2+...)*u and F(w) = g_1(w)*L1+g_2(w)*L2+... with
-    parameter w_i in [a_i,b_i], where f_i(w) and g_i(w) are strings of valid python
-    expressions.
-
-    Optional products can be provided to introduce a dict of inner products on
-    the discrete space. The content of the file is then given as:
-
-    [stiffness-matrices]
-    # path_to_object: parameter_functional_associated_with_object
-    K1.mat: f_1(w_1,...,w_n)
-    K2.mat: f_2(w_1,...,w_n)
-    ...
-    [rhs-vectors]
-    L1.mat: g_1(w_1,...,w_n)
-    L2.mat: g_2(w_1,...,w_n)
-    ...
-    [parameter]
-    # Name: lower_bound,upper_bound
-    w_1: a_1,b_1
-    ...
-    w_n: a_n,b_n
-    [products]
-    # Name: path_to_object
-    Prod1: S.mat
-    Prod2: T.mat
-    ...
-    """
-    assert ".ini" == parameter_file[-4:], "Given file is not an .ini file"
-
-    # Get input from parameter file
-    config = ConfigParser.ConfigParser()
-    config.optionxform = str
-    config.read(parameter_file)
-
-    # Assert that all needed entries given
-    assert 'stiffness-matrices' in config.sections()
-    assert 'rhs-vectors' in config.sections()
-    assert 'parameter' in config.sections()
-
-    stiff_mat        = config.items('stiffness-matrices')
-    rhs_vec          = config.items('rhs-vectors')
-    parameter        = config.items('parameter')
-
-    # Dict of parameters types and ranges
-    parameter_type = {}
-    parameter_range = {}
-
-    # get parameters
-    for i in range(len(parameter)):
-        parameter_name = parameter[i][0]
-        parameter_list = tuple(float(j) for j in parameter[i][1].replace(" ","").split(','))
-        parameter_range[parameter_name] = parameter_list
-        # Assume scalar parameter dependence
-        parameter_type[parameter_name] = 0
-
-    # Create parameter space
-    parameter_space = CubicParameterSpace(parameter_type=parameter_type,ranges=parameter_range)
-
-    # Assemble operators
-    stiff_operators, stiff_functionals = [], []
-
-    # get parameter functionals and stiffness matrices
-    for i in range(len(stiff_mat)):
-        path = stiff_mat[i][0]
-        expr = stiff_mat[i][1]
-        parameter_functional = ExpressionParameterFunctional(expr, parameter_type=parameter_type)
-        info = io.loadmat(path, mat_dtype=True).values()
-        stiff_operators.append(NumpyMatrixOperator(info[[j for j in range(len(info)) if type(info[j]) == np.ndarray][0]]))
-        stiff_functionals.append(parameter_functional)
-
-    stiff_lincombOperator = LincombOperator(stiff_operators, coefficients=stiff_functionals)
-
-    # get rhs vectors
-    rhs_operators, rhs_functionals = [], []
-
-    for i in range(len(rhs_vec)):
-        path = rhs_vec[i][0]
-        expr = rhs_vec[i][1]
-        parameter_functional = ExpressionParameterFunctional(expr, parameter_type=parameter_type)
-        info = io.loadmat(path, mat_dtype=True).values()
-        rhs_operators.append(NumpyMatrixOperator(info[[j for j in range(len(info)) if type(info[j]) == np.ndarray][0]].T))
-        rhs_functionals.append(parameter_functional)
-
-    rhs_lincombOperator = LincombOperator(rhs_operators, coefficients=rhs_functionals)
-
-    # get products if given
-    if 'products' in config.sections():
-        product = config.items('products')
-        products = {}
-        for i in range(len(product)):
-            product_name = product[i][0]
-            product_path = product[i][1]
-            info=io.loadmat(product_path,mat_dtype=True).values()
-            products[product_name] = NumpyMatrixOperator(info[[j for j in range(len(info)) if type(info[j]) == np.ndarray][0]])
-    else:
-        products = None
-
-    # Create and return stationary discretization
-    return StationaryDiscretization(operator=stiff_lincombOperator, rhs=rhs_lincombOperator, parameter_space=parameter_space, products=products)
